requirements.txt
fastapi[standard]==v0.115.12
uvicorn[standard]==0.34.3
sqlalchemy==2.0.41
python-dotenv==1.1.0
psycopg2-binary==2.9.10
pydantic==2.11.5
